crawlers/requirements.txt
# Web scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# For JavaScript-rendered pages
playwright>=1.40.0

# RSS/Atom feed parsing
feedparser>=6.0.0

# Date parsing
python-dateutil>=2.8.0

# HTML to text
html2text>=2020.1.16

# Rate limiting
ratelimit>=2.2.1

# Async support
aiohttp>=3.9.0

# Environment variables
python-dotenv>=1.0.0
